allow create/delete[Function|Device] to take different argument types

+++ src/Hooks/useDevices.ts
@@ -1,6 +1,7 @@
 import {useCallback, useLayoutEffect, useState} from 'react';
 import {useGlobalLynxClient} from '../Contexts';
-import {Devicex, EmptyDevicex, ErrorResponse, Metadata} from '@iotopen/node-lynx';
+import {Devicex, EmptyDevicex, ErrorResponse, Metadata, OKResponse} from '@iotopen/node-lynx';
+import {ObjectOrArray} from '../types';
 
 export const useDevices = (installationId: number, filter?: Metadata) => {
     const {lynxClient} = useGlobalLynxClient();
@@ -17,40 +18,50 @@ export const useDevices = (installationId: number, filter?: Metadata) => {
         }).finally(() => {
             setLoading(false);
         });
-        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [lynxClient, installationId, filter]);
 
-    const remove = useCallback((dev: Devicex[]) => {
-        const last = dev.pop();
-        if (!last) return Promise.allSettled([]);
-        const rest = dev.map((f => {
-            return lynxClient.deleteDevice(f, true);
-        }));
-        return Promise.allSettled(rest).then(async (settled) => {
-            try {
-                settled.push({status: 'fulfilled', value: await lynxClient.deleteDevice(last)});
-            } catch (e) {
-                settled.push({status: 'rejected', reason: e});
-            }
-            return settled;
-        });
-    }, [lynxClient]);
+    function removeFn<T extends Devicex | Devicex[]>(devs: T): ObjectOrArray<OKResponse, Devicex, T>
+    function removeFn(devs: Devicex | Devicex[]) {
+        if (Array.isArray(devs)) {
+            const last = devs.pop();
+            if (!last) return Promise.allSettled([]);
+            const rest = devs.map((dev => {
+                return lynxClient.deleteDevice(dev, true);
+            }));
+            return Promise.allSettled(rest).then(async (settled) => {
+                try {
+                    settled.push({status: 'fulfilled', value: await lynxClient.deleteDevice(last)});
+                } catch (e) {
+                    settled.push({status: 'rejected', reason: e});
+                }
+                return settled;
+            });
+        }
+        return lynxClient.deleteDevice(devs);
+    }
 
-    const create = useCallback((dev: EmptyDevicex[]) => {
-        const last = dev.pop();
-        if (!last) return Promise.allSettled([]);
-        const rest = dev.map(f => {
-            return lynxClient.createFunction(f, true);
-        });
-        return Promise.allSettled(rest).then(async (settled) => {
-            try {
-                settled.push({status: 'fulfilled', value: await lynxClient.createDevice(last)});
-            } catch (e) {
-                settled.push({status: 'rejected', reason: e});
-            }
-            return settled;
-        });
-    }, [lynxClient]);
+    function createFn<T extends EmptyDevicex | EmptyDevicex[]>(devs: T): ObjectOrArray<Devicex, EmptyDevicex, T>
+    function createFn(devs: EmptyDevicex | EmptyDevicex[]) {
+        if (Array.isArray(devs)) {
+            const last = devs.pop();
+            if (!last) return Promise.allSettled([]);
+            const rest = devs.map(dev => {
+                return lynxClient.createDevice(dev, true);
+            });
+            return Promise.allSettled(rest).then(async (settled) => {
+                try {
+                    settled.push({status: 'fulfilled', value: await lynxClient.createDevice(last)});
+                } catch (e) {
+                    settled.push({status: 'rejected', reason: e});
+                }
+                return settled;
+            });
+        }
+        return lynxClient.createDevice(devs);
+    }
+
+    const create = useCallback(createFn, [lynxClient]);
+    const remove = useCallback(removeFn, [lynxClient]);
 
     useLayoutEffect(() => {
         refreshCall();
