+++ src/Hooks/useFunctions.ts
@@ -1,6 +1,7 @@
 import {useCallback, useLayoutEffect, useState} from 'react';
 import {useGlobalLynxClient} from '../Contexts';
-import {EmptyFunctionx, ErrorResponse, Functionx, Metadata, MetaObject} from '@iotopen/node-lynx';
+import {EmptyFunctionx, ErrorResponse, Functionx, Metadata, OKResponse} from '@iotopen/node-lynx';
+import {ObjectOrArray} from '../types';
 
 export const useFunctions = (installationId: number, filter?: Metadata) => {
     const {lynxClient} = useGlobalLynxClient();
@@ -17,40 +18,50 @@ export const useFunctions = (installationId: number, filter?: Metadata) => {
         }).finally(() => {
             setLoading(false);
         });
-        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [lynxClient, installationId, filter]);
 
-    const remove = useCallback((fn: Functionx[]) => {
-        const last = fn.pop();
-        if (!last) return Promise.allSettled([]);
-        const rest = fn.map((f => {
-            return lynxClient.deleteFunction(f, true);
-        }));
-        return Promise.allSettled(rest).then(async (settled) => {
-            try {
-                settled.push({status: 'fulfilled', value: await lynxClient.deleteFunction(last)});
-            } catch (e) {
-                settled.push({status: 'rejected', reason: e});
-            }
-            return settled;
-        });
-    }, [lynxClient]);
+    function removeFn<T extends Functionx | Functionx[]>(fns: T): ObjectOrArray<OKResponse, Functionx, T>
+    function removeFn(fns: Functionx | Functionx[]) {
+        if (Array.isArray(fns)) {
+            const last = fns.pop();
+            if (!last) return Promise.allSettled([]);
+            const rest = fns.map((f => {
+                return lynxClient.deleteFunction(f, true);
+            }));
+            return Promise.allSettled(rest).then(async (settled) => {
+                try {
+                    settled.push({status: 'fulfilled', value: await lynxClient.deleteFunction(last)});
+                } catch (e) {
+                    settled.push({status: 'rejected', reason: e});
+                }
+                return settled;
+            });
+        }
+        return lynxClient.deleteFunction(fns);
+    }
 
-    const create = useCallback((fn: EmptyFunctionx[]) => {
-        const last = fn.pop();
-        if (!last) return Promise.allSettled([]);
-        const rest = fn.map(f => {
-            return lynxClient.createFunction(f, true);
-        });
-        return Promise.allSettled(rest).then(async (settled) => {
-            try {
-                settled.push({status: 'fulfilled', value: await lynxClient.createFunction(last)});
-            } catch (e) {
-                settled.push({status: 'rejected', reason: e});
-            }
-            return settled;
-        });
-    }, [lynxClient]);
+    function createFn<T extends EmptyFunctionx | EmptyFunctionx[]>(fns: T): ObjectOrArray<Functionx, EmptyFunctionx, T>
+    function createFn(fns: EmptyFunctionx | EmptyFunctionx[]) {
+        if (Array.isArray(fns)) {
+            const last = fns.pop();
+            if (!last) return Promise.allSettled([]);
+            const rest = fns.map(f => {
+                return lynxClient.createFunction(f, true);
+            });
+            return Promise.allSettled(rest).then(async (settled) => {
+                try {
+                    settled.push({status: 'fulfilled', value: await lynxClient.createFunction(last)});
+                } catch (e) {
+                    settled.push({status: 'rejected', reason: e});
+                }
+                return settled;
+            });
+        }
+        return lynxClient.createFunction(fns);
+    }
+
+    const create = useCallback(createFn, [lynxClient]);
+    const remove = useCallback(removeFn, [lynxClient]);
 
     useLayoutEffect(() => {
         refreshCall();
